Reixed bug 3
Changed approach to fix bug 3 since it introduced new bugs. New approach changes only one variable which is to assume player made false move in the beginning. If player is assumed to make false move before even playing a move, then the function responsible to lead the player to the next move such as incrementing the player move count or making new move will be ignored